Opprydding og refaktorering etter code review
