+++ src/main/kotlin/no/nav/familie/ef/søknad/api/OppslagController.kt
@@ -3,7 +3,7 @@ package no.nav.familie.ef.søknad.api
 
 import no.nav.familie.ef.søknad.api.dto.Søkerinfo
 import no.nav.familie.ef.søknad.service.Oppslag
-import no.nav.familie.ef.søknad.util.TokenUtil.Companion.ISSUER
+import no.nav.familie.ef.søknad.util.InnloggingUtils
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.slf4j.LoggerFactory
 import org.springframework.http.MediaType.APPLICATION_JSON_VALUE
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping(path = [OppslagController.OPPSLAG], produces = [APPLICATION_JSON_VALUE])
-@ProtectedWithClaims(issuer = ISSUER, claimMap = ["acr=Level4"])
+@ProtectedWithClaims(issuer = InnloggingUtils.ISSUER, claimMap = ["acr=Level4"])
 class OppslagController(private val oppslag: Oppslag) {
 
     private val log = LoggerFactory.getLogger(OppslagController::class.java)
