+++ src/main/kotlin/no/nav/familie/ef/søknad/api/PingController.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ef.søknad.api
 
 import no.nav.familie.ef.søknad.api.dto.Søknad
-import no.nav.familie.ef.søknad.util.TokenUtil
+import no.nav.familie.ef.søknad.util.InnloggingUtils
 import no.nav.security.token.support.core.api.Protected
 
 import org.springframework.http.MediaType
@@ -19,13 +19,13 @@ class PingController {
 
     @PostMapping("/ping", consumes = [MediaType.APPLICATION_JSON_VALUE])
     fun pingPost(@RequestBody request: Søknad): String {
-        return " Ack - vi har mottatt: ${request.text} ${TokenUtil().fødselsnummer}"
+        return " Ack - vi har mottatt: ${request.text} ${InnloggingUtils.fødselsnummer}"
     }
 
 
     @GetMapping("/getToken")
     fun getToken(): String {
-        return TokenUtil().fødselsnummer
+        return InnloggingUtils.fødselsnummer
     }
 
 }
