+++ src/main/kotlin/no/nav/familie/ef/søknad/config/ApplicationConfig.kt
@@ -28,7 +28,7 @@ class ApplicationConfig(@Value("\${application.name}") val applicationName: Stri
 
     @Bean
     fun corsFilter(corsProperties: CorsProperties): FilterRegistrationBean<CORSResponseFilter> {
-        LOG.info("Registering LogFilter filter")
+        LOG.info("Registrerer CORSResponseFilter")
         val filterRegistration = FilterRegistrationBean<CORSResponseFilter>()
         filterRegistration.filter = CORSResponseFilter(corsProperties)
         filterRegistration.order = 0
