+++ src/main/kotlin/no/nav/familie/ef/søknad/config/FeatureToggleConfig.kt
@@ -57,7 +57,6 @@ class FeatureToggleConfig(@Value("\${familie.ef.funksjonsbrytere.enabled}") val 
     private fun lagDummyFeatureToggleService(): FeatureToggleService {
         return object : FeatureToggleService {
             override fun isEnabled(toggleId: String, defaultValue: Boolean): Boolean {
-                return defaultValue
                 if(unleashEnv == "local"){
                     return true;
                 }
