+++ src/main/kotlin/no/nav/familie/ef/søknad/service/mellomlagring/StorageService.kt
@@ -2,25 +2,24 @@ package no.nav.familie.ef.søknad.service.mellomlagring
 
 import no.nav.familie.ef.søknad.exception.AttachmentVirusException
 import no.nav.familie.ef.søknad.service.VirusScanner
-import no.nav.familie.ef.søknad.util.TokenUtil
+import no.nav.familie.ef.søknad.util.InnloggingUtils
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 
 @Service
-class StorageService(private val tokenHelper: TokenUtil,
-                     private val storage: Storage,
+class StorageService(private val storage: Storage,
                      private val virusScanner: VirusScanner,
                      private val crypto: StorageCrypto) {
 
     fun hentSøknad(): String? {
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.trace("Henter søknad fra katalog {}", directory)
         return storage.getTmp(directory, SØKNAD)?.let { crypto.decrypt(it, fnr) }
     }
 
     fun lagreSøknad(søknad: String) {
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.trace("Skriver søknad til katalog {}", directory)
         val encryptedValue = crypto.encrypt(søknad, fnr)
@@ -28,14 +27,14 @@ class StorageService(private val tokenHelper: TokenUtil,
     }
 
     fun slettSøknad() {
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.info("Fjerner søknad fra katalog {}", directory)
         storage.deleteTmp(directory, SØKNAD)
     }
 
     fun hentVedlegg(key: String): Vedlegg? {
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.info("Henter vedlegg med nøkkel {} fra katalog {}", key, directory)
         return storage.getTmp(directory, key)?.let { Vedlegg.fromJson(crypto.decrypt(it, fnr)) }
@@ -45,7 +44,7 @@ class StorageService(private val tokenHelper: TokenUtil,
         if (!virusScanner.scan(vedlegg)) {
             throw AttachmentVirusException(vedlegg)
         }
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.info("Skriver vedlegg {} til katalog {}", vedlegg, directory)
         val encryptedValue = crypto.encrypt(vedlegg.toJson(), fnr)
@@ -53,13 +52,13 @@ class StorageService(private val tokenHelper: TokenUtil,
     }
 
     fun slettVedlegg(key: String) {
-        val directory = crypto.encryptDirectoryName(tokenHelper.fødselsnummer)
+        val directory = crypto.encryptDirectoryName(InnloggingUtils.fødselsnummer)
         LOG.info("Fjerner vedlegg med nøkkel {} fra katalog {}", key, directory)
         storage.deleteTmp(directory, key)
     }
 
     fun hentKvittering(type: String): String? {
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.trace("Henter kvittering fra katalog {}", directory)
 
@@ -68,7 +67,7 @@ class StorageService(private val tokenHelper: TokenUtil,
 
 
     fun lagreKvittering(type: String, kvittering: String) {
-        val fnr = tokenHelper.fødselsnummer
+        val fnr = InnloggingUtils.fødselsnummer
         val directory = crypto.encryptDirectoryName(fnr)
         LOG.trace("Skriver kvittering til katalog {}", directory)
         val encryptedValue = crypto.encrypt(kvittering, fnr)
