+++ src/test/resources/logback-spring-test.xml
@@ -6,5 +6,4 @@
               value="%d [%-5level] [%thread] %logger{5} %replace(- [%X{consumerId}, %X{callId}, %X{userId}] ){'- \[, , \] ',''}- %m%n"/>
     <include resource="org/springframework/boot/logging/logback/base.xml"/>
 
-    <logger name="no.nav.security" level="DEBUG"/>
 </configuration>
